pyproject.toml
[project]
name = "comparison-of-optimizers"
version = "0.1.0"
python = ">=3.12,<3.13"
package-mode = false
dependencies = ["torch>=2.2.0", "torchvision>=0.17.0", "lion-pytorch>=0.1.0", "matplotlib>=3.8.0", "pandas>=2.2.0", "numpy>=1.26.0", "pyyaml>=6.0.1", "omegaconf>=2.3.0", "types-PyYAML>=6.0.12"]


[project.optional-dependencies]
dev = ["flake8>=7.0.0", "mypy>=1.9.0", "black>=25.9.0", "pre-commit>=4.3.0", "tomli"]

[tool.black]
line-length = 99
target-version = ['py312']

[tool.mypy]
files = ["src/"]
python_version = "3.12"
strict = false
ignore_missing_imports = true
disable_error_code = ["no-untyped-def", "type-arg", "no-untyped-call"]

[[tool.mypy.overrides]]
module = ["omegaconf.*"]

[tool.flake8]
max-line-length = 99
extend-ignore = "E203, W503"
